MB-41321: Combine Collections related objects 'CommitData' and 'Flush'
The CommitData which is only a member of KVStore is now moved and
combined into Collections:VB::Flush. This is required by MB-41321 so we
can more easily check if a flushed key is live/dropped in a flush batch
and also cleans up some of the code managing/recording flushing of
system-events (no need for KVStore to call clear etc...).

Part of this merging also moves some of the persisted 'seqno' members to
be unsigned types - we never persist a negative seqno and this helps
reduce the mental burden in later code where are comparing sequence
numbers from couchstore/magma with the Flush object,

Change-Id: Id7e1925d5520949fa874c1a384670cbed6f61821