MB-16181: Restructure Collections::Manifest
The class has never been great since the addition of scopes with
some data in the collection map and some data in
the scopes map. This commit updates the object so that there is
one 'CollectionEntry' that holds all collection meta-data,
the collection map now just references that one entry.

Change-Id: I547da9020c982804c803c673f3e610c6756a7207