MB-41321: 2/4 Clean-up if collection exists in old and new drop containers
Collections::Flush was erasing a collection from the new drop 'map'
when it was found to be in both 'old' and 'new' - this is so the
flatbuffer output (which is a vector) only had the dropped collection
once.

However this leads to code executing later in the flush to not see
the collection when iterating through the 'droppecCollections' map.
For example notifyManifestOfAnyDroppedCollections would not get the
VB::Manifest to clean-up (effectively leaking memory).

This fix ensures we still generate the collection once in the
flatbuffers output, but also keep it in the list and now clean-up.

Change-Id: I4f05d8728bb169bce309c5331ed41fb41c1a7ecd