MB-41882: Detect an empty collection being dropped and skip purge
An empty collection is generally one where the high-persisted-seqno
equals the collection's start-seqno, although a high-persisted-seqno
of 0 also indicates empty for a create/drop in a single flush.

Using the MB-41321 changes it is now much easier to detect if
an empty collection has been dropped, KV-engine should skip
scheduling a purge when this happens.

Change-Id: Iba1e6840c8d157fe30fed3183f381743f7723b85