MB-39292: Allow a forced collection update to reset the manifest uid
The following cases of update are currently invalid and will be enabled
by force in this commit.

* Update where the manifest UID goes backwards
* Update where the manifest UID is equal
* Update where the manifest UID goes forwards, but makes no change

With this commit when a manifest is pushed to KV with force=true
any detectable changes will be applied. Detectable changes currently
only means when scope/collections are added or dropped.

In the case of a UID change, but no changes to collections this propagates
to VB::Manifest as a UID only change.

Change-Id: I1962ffe04f1fe6b46aab80b54410247444cc8483