MB-39292: 1/n VB::Manifest holds shared_ptr to the manager
This is 1 of a series of patches that improve collection metadata
management with the final goal of allowing collection state changes
to detect scope/collection name changes (when forced). This can
happen as part of a set-collections command or during vbucket state
change (explicit command or DCP takeover both need this).

Some of the aims of this series:

* To keep vbucket state changes reasonably similar in behaviour to pre
  7.0 the name data needs to be accessible reasonably swiftly (so not
 dependent on disk reads).

* As collections/scopes generally have the same name across nodes and
  vbuckets and that a could be 251 bytes do not keep a copy of the name
  in each collection/scope entry.

This commit tweaks the ownership of the KBucket 'owned'
Collections::Manager, which is now a shared_ptr.

Next the VB::Manifest object is updates so it also has a shared_ptr
for its Manager class (so that later we can access/update shared names).

The use of shared_ptr reflects the fact that a VBucket can outlive the
KVBucket it logically belongs to, this primarily happens when the deferred
VBucket delete task runs after the KVBucket has been deleted.

Change-Id: Ib037cb178848fe9cdcb8b3dbe7ae7267738f0f3d