MB-39292: 2/n Add a 'SharedMetaDataTable'
To achieve the goal of this series of patches, we need to access
the name of a collection so we can correctly process a force
re-create.

To avoid increasing memory usage, we need a place in memory where
the VB::ManifestEntry objects will be able to reference their name.
The stored data should be store once per bucket and not duplicated
for all vbuckets, because names can be upto 251 bytes.

This patch introduces a container type for the storage and retrieval
of meta data to be associated with collections (and scopes).

The intention is that the name, scope-id and ttl of a collection must
not be stored in the VB::ManifestEntry, but referenced (and
ref-counted). This container deliberately does not use shared_ptr
because we want to avoid any further per collection memory increase. So
the container gives out a reference to the stored meta and later users
revoke their reference (with dereference calls). The container can
then erase values from the internal map once 0 users have a reference.

The container uses a multimap for storage because we need to support
different id->meta mappings. E.g.

* vbucket 0 has collection:x with name:fruit
* vbucket 1 has collection:x with name:vegetable

This can happen in failure circumstances when the vbuckets see
different state time-lines (which will be force corrected).

Note the SharedMetaDataTable will be used by VB::Manifest for scope
name storage too, hence the full templating of all key and value.

Change-Id: I7c7ded00827bf4fa1978668a1ffaebe908749dc6