MB-45221: 1/n Collection stats delete->delete incorrectly handled
When KVStore performs an update of a deleted item with another,
the Collections::Flush::updateStats function incorrectly treats
the change as an insert, meaning the diskSize increased by the
size of the new item, rather than by the difference of old vs new.

As part of the fix, aborts are now included in the collection disk
size, this simplifies the tracking by removing the "if abort"
special cases.

The fix restructures the stat update code to make the various states
much more clear. The following states are considered when we are
updating a key in the KVStore (this is also duplicated in the code as
a comment).

The following defines our expected old and new states. Note that only
committed items actually increment the item count (and that is handled
in Flush::StatisticsUpdate::insert/remove).

The old key can be live, deleted or dropped.
The new key can be live or deleted.

new key is live:
  * old key dropped: Key is an insert => items += 1, diskSize += size
  * old key deleted: Key is an insert => items += 1, diskSize += delta
  * old key live: Key is an update => diskSize += delta

new key is deleted
  * old key is dropped: Key is an update => diskSize += size
  * old key is deleted: Key is an update => diskSize += delta
  * old key is live: Key is a remove => items -= 1, diskSize += delta

Note that old can be both dropped and deleted (a dropped tombstone). In
that case we process as dropped.

Change-Id: I21e02085dd903fab416117c9b2be348c8bc9ccde