MB-45504: Add HistoryID class and tests
Add a class which can store a HistoryID.

The history ID will be transmitted to KV as a 32 character hex-string
representing a 128-bit UUID.

KV will need to store the ID in a number of places (including in
flatbuffer persisted data). The HistoryID class wraps a flatbuffer
type and allows construction from the 32 character string or an
existing flatbuffer type. The class can convert back to string or
flatbuffer type.

Change-Id: I5770fee565cc636bc77cd705f2c93ed83409d31a