MB-45505: Collections::Manifest support history_id
Update Collections::Manifest so that it now requires a
"history_id" in the JSON manifest. Note that a default
constructor remains due to the use of folly::Synchronized
in the Manager class.

This commit mostly removes the force flag support and this
results in a number of test no longer working.

The test which no longer work are now disabled as they
require the use of the force flag, but a number of
these tests are incorrect tests that will be fixed
or removed once we have history_id wired further through
the collection update code.

Change-Id: I458deece70e6efd61079ad5bfe2e01b9acbe469b