MB-45505: VB::Filter 'uid' clean-up
Remove the uid parsing and optional member along with some dead
code from VB::Filter.

The uid serves no current purpose other than getting logged.
Removing this now to simplify later changes related to MB-45505.

Commit is updated over a previous version with a unit test that
demonstrates a filter with the "uid" key is still valid, we just
don't parse it any and store it.

Change-Id: If3cbd326d780755b00c999b82e9f46373dffd24f