MB-49040: 1/n Track the scope's data usage when flushing
VB::Manifest now maps from scope-id to a VB::ScopeEntry that stores
all about a scope, that is the (new) data-size and the (existing) name.

As the flusher updates the collection data-size, a new call to
updateScopeDataSize will find the collection's VB::ScopeEntry and
update the counter.

Note: With this commit warmup will reset the scope data-size to 0, a
later commit will warmup and set the initial size.

Change-Id: I2b7c05e0b06975a28f4b57b27fea97cfa11cc98e