MB-49040: 4/n Modify a scope when the Manifest is different
As vbuckets switch to active they always check:

Q: Does the vbucket have all scopes and collections from the
   Manifest?

Done by comparing Collections::VB::Manifest with Collections::Manifest.
The outcome of that compare results in creations and drops to make
them equal.

This check is now updated to include a compare of the data-limit,
even when sets of scopes/collections are equal.

Q: Do vbucket scopes have the same data-limit *value* as the Manifest?

This check is needed because data-limits are not replicated in scope
creation events - so a replica VB::Manifest will not directly be given
the limit when it is created via 'replicaCreateScope', instead when
a vbucket becomes switched from replica to active, at that point we
will adjust the data limit.

Note the scope DataLimit is stored in the ScopeSharedMetaData. This
means that provided a scope with limit is first created on an active
vbucket the ScopeSharedMetaData will store the correct data limit. So
only when a scope with limit is first created on a replica will the
modification be required, and this happens for example when a new node
is added to a cluster and all data, collections and scopes are created
by replication.

Change-Id: Ic155e1a917d38c1b084d70be4ead9efac008ae60