TEST ME MB-49040: 5/n Update scope data-size during warmup
Update some existing tests and see that the data size comes
back on warmup
...

Establish the baseline data-size at warmup during the population
of collection stats from the store statistics. This is done by
passing the collection size through to the manifest so it can
find and update the collection's scope.

Change-Id: Ibc2892ec2a1450befb7cada37573b457b0aaeff4