MB-35297: Initial RangeScan infrastructure
This commit adds very basic RangeScan support and supporting tests.

A RangeScan object is defined which is an object that is intended to
be created in response to a range-scan request command. The object
itself will then be driven through various states by a mixture of
I/O tasks (opening and scanning disk) and frontend threads requesting
create/continue/cancel. The RangeScan object is given a number of
parameters, primarily the "handler" which abstracts the end point
for keys and values that are read in the scan. Test code defines
a handler that just stored key/values in a vector.

A RangeScanOwner object is defined that is the owner of all scans
against a vbucket. The RangeScanOwner object provides methods that
move a scan through its lifetime state - Idle, Continue, Cancel

The code in range_scan_test shows the current abilities of this commit,
configuring scans for various ranges (key or value) and also checking a
number of invariants regarding the ordering of client operations. There
are no limits defined so each scan returns the entire range in one
continue.

Change-Id: I694dee94902bcaa13dd4d509da9a9edc5cb94148