MB-35297: Return backfill_finished instead of transition to done
If the backfill_status is backfill_finished then the task will no
longer be executed, so transitionState has no effect and complicates
the create/scan functions.

This appears a more efficient approach, certainly for the cases where
success used to be returned. The task is now removed from backfill
manager sooner and requires 1 less invocation to reach
backfill_finished.

Change-Id: I1e4f44eb3cee5535302296dbf79e70d765111032