MB-35297: Remove complete from the DCPBackfill::cancel path
DCPBackfill::cancel is invoked when shouldCancel returns true.
This only happens if:

* ActiveStream is dead
* ActiveStream has gone (weak_ptr cannot lock)

In either of those cases the complete function (implemented at the
BySeqno/ByID) level of the backfill should not be called.

1) If the ActiveStream has gone, complete will fail to lock the
   weak_ptr.
2) If the stream can be obtained, calling completeBackfill (or OSO
   equivalent) is pointless on a dead stream, for example any
   DcpResponse queued is going no where.

This commit aims to simplify based on the above description, removing
the call to complete from the cancel path. This makes the
implementation of complete much simpler as it doesn't need to consider
logging "cancel".

Change-Id: I01e786ada63888f46ac5eac01fd683b60d6f6717