MB-35297: Don't use cb::engine_errc::no_mem when a scan needs to yield
Scans could need to yield for a variety of reasons, not just because
of a buffer being full (no_mem). This commit adds a yield/shouldYield
pair of methods to the StatusCallback to make it clearer when code
wants to pause/yield.

Underneath the yield method is a change to use 'temporary_failure'
as the status code.

Change-Id: If2260b3ed475d1441e8e8db1677e6809f9390f48