MB-35297: Continue/Cancel RangeScans on a I/O task
Add RangeScanContinueTask (READER) which is a task to be scheduled
in response to continue or cancel requests.

For continue, the "scanning" occurs on the I/O task. For cancel, the
destruct (and close of snapshot) occurs on the same I/O task

In this commit the task is quite simple primarily to build code around
continuing without too much complexity in the task itself. In this
commit one new task is created and scheduled per continue/cancel, which
is adequate for now, future changes will look to limit number of tasks and
concurrency.

Change-Id: I96daac2c6b81da145f21754e17c519de13cafc03