MB-35297: Add an item limit to range scan continue
Update the continueRangeScan API so an item limit can be set.
This limit is pushed down to the RangeScan object and then
checked as keys/items are 'read' by the scan. Hitting the limit
means the KVStore::scan is told to yield and the client will
then need to re-continue the scan when ready.

Change-Id: I4690d5fa2d7812bfe2e6a0a214a115163a212550