MB-35297: Add optional random sampling mode to range-scan
The range scan functionality can be adapted so that a client
could request that the scan returns a random number of keys
or documents.

This commit adds an optional configuration parameter that
will allow a client to set a prng seed value and a number
of samples to return. If the collection in the snapshot
has more items than samples, the range scan continue will
return a random set of keys or documents.

Change-Id: I3ca3af43e5e04d621a4f1df164fefcaebd106149