MB-35297: Add "wait-for-persistence" to rangeScanCreate
Add to the range-scan-create path a "wait-for-persistence" for
feature which allows the clients creation to specify a uuid/seqno
and timeout.

If the vbucket uuid matches and the seqno hasn't been persisted
the command will wait for the timeout (using ewouldblock).

This now means that the range-scan-create path now can perform two
ewouldblock steps.

1) The wait for persistence to be satisfied and notify
2) The wait for RangeScanCreateTask to run and notify

The state a in-progress range-scan-create is in is now captured in the
state variable which is stored in the engine-specific.

Pending is the initial state
WaitForPersistence waiting for SeqnoPersistenceRequest
Create waiting for RangeScanCreateTask

Change-Id: I100ab255b6a46ca85c807a0f38c132c81bab936a