MB-35297: Fail RangeScanCreate if a range has no data
Instead of allowing a scan to exist which will return nothing
when continued, detect an empty range at the create stage and
fail the request if the range is empty.

This commit achieves this by running the scan for 1 key only and
detecting that one key was 'loaded' by the status returned from
KVStore::scan.

If an empty range is requested, the create command returns to
the client "KeyEnoent" (status 0x1).

Change-Id: I6ca4572f7accdadab493ba19860df7466b263e2c