MB-35297: RangeScan continue using mcbp
Add mcbp validator/executor so that an existing RangeScan can be
continued.

A RangeScan continue is a bit different to other mcbp commands.

1) A single continue request may trigger many response packets, with
each response carrying a payload containing many keys/values.
See range_scan_continue.md for a description of the payload format.

2) The RangeScan I/O task (AUXIO) is writing 'directly' to the
connection. This is different to the usual ewould_block pattern
as we don't bring data into memory and then trigger a worker thread
to copy that memory out. Note that the continue request does not
use out-of-order and the connection cannot process any more commands,
even whilst waiting for the AUXIO task to pickup the continue task.

Change-Id: I3dcd0c8445190bb1f313d2b2746779bff7ff3cbf