MB-35297: Cancel RangeScans that exist for "too long"
A RangeScan cannot 'exist' indefinitely as disk cannot be released by
compaction. This commit adds a task that will 'force' cancel any
RangeScan that has existed for longer than a configuration defined
duration (180 seconds by default). This cancelling occurs irrespective
of the scan state/progress.

A cancellation task is created per vbucket and shares code with the
durability timeout task. The idea is that when a scan is created, the
task will change from sleep(-1) to sleep(180). When the task wakes it
will check the state of the vbucket's scans and force cancel any that
have a 180 second existence. At this point the task may sleep(-1) or
set a sleep time based on the vbucket's scans, e.g. if scans remain
the sleep time is set to be the duration until the next scan would
of exceeded the 180 seconds.

Change-Id: I409b00ab08772fa36af8f2fad83c0b083fe26165