MB-52957: Avoid scan when collection high seqno < start
For filtered DCP streams check if the high seqno of the scanned
collections is lower than the start seqno of the scan.

If this check returns true, no documents for the collection(s) will be
found by scanning. Prior to this change we would scan the seqno index
from start to end, but produce no data for the DCP client except for
an empty snapshot+seqno-advance.

Note: some ep_testsuite_dcp tests are now skipped under RocksDB as it
lacks collection support. Those ep_testsuite_dcp tests are all acting
like legacy DCP client - i.e. filtering for the default collection
and thus trigger the early finish of the backfill as the high-seqno
of the default collection is incorrectly reported as 0.

Change-Id: Ifda930fd6fd2d1981a24a04724e0146bcf3062c3