MB-52956: Update lastReadSeqno at the end of an OSO backfill
The idea behind an OSO backfill is to open a snapshot that stores many
collections, yet read only the part of the index the encompasses the
desired collection - all done using the key index.

Once the OSO snapshot has finished, generally we can then switch to
in-memory streaming and pickup collection changes from memory.

The MB identified that if the cursor was dropped during the OSO
backfill, the next backfill starts too early.

E.g.

Snapshot 1, the OSO backfill was a snapshot for seqnos 0 to 1000 and we
read 1 collection item at seqno 5.

The next backfill runs to fill in the gaps and the next snapshot
represents seqnos 0 to 1500, ActiveStream starts the backfill at seqno
6, which is incorrect. The first OSO snapshot has effectively skipped 6
to 1000, we know that range stores no data for our collection.

By updating lastReadSeqno to maxSeqno in the above example the second
backfill begins at 1001, reducing the scan size to the correct range.

Change-Id: If65bfa3c59310ca707ac29f53962958f35745c9e