MB-53428: Disable and handle any attempt renegotiation SSL
1) Disable OpenSSL renegotiation

This is done by adding SSL_OP_NO_RENEGOTIATION to SSL_CTX_set_options.

2) Ensure KV responds to a renegotiation attempt by checking if
SSL_read has generated a response.

With renegotiation disabled, a client can still attempt to renegotiate.
When this happens an SSL alert is generated by the server and must be
sent back to the client. This is achieved by querying the network BIO
after calling SSL_read if the WANT_READ error occurs. When write data
has been generated call the data is written back to the client and
the read retried.

2.1) Read and process SSL input until connection ends

As part of testing the renegotiation an inconsistency between
mad-hatter and cheshire-cat was noted whilst using the
'openssl s_client' for testing.

The flow of messages was observed in wireshark as follows.

client -> server: Handshake
server -> client: SSL alert, no_renegotiation
client -> server: SSL alert, no_renegotiation

I.e. the client appears to echo back the SSL alert, the openssl client
would also terminate immediately after sending the last message. KV had
a racey behaviour in that sometimes KV engine would process the incoming
alert (and reset the connection) and sometimes it would not see the
message, compared to cheshire-cat where KV would always process the
incoming alert.

This is addressed by moving the check of ssl.hasError to be always after
the sslRead function. When hasError was before the sslRead call, sometimes
we would skip processing the final bytes received from the client.

Change-Id: I5517d4558142d4b6616be86953f7261d9419593c