MB-35297: Move BackfillTrackingIface out of DcpConnMap
This change is a refactoring change that forms part of the overall
change which will ensure RangeScans are capped in terms of how many
file-descriptors are in-use.

The current BackfillTrackingIface which is implemented in DcpConnMap
is used to count and limit how many DCP backfills are running, ensuring
DCP doesn't just keep opening files and consuming all file-descriptors.

RangeScans will also participate in this calculation and will need to
call into this object.

The refactoring:

1) moves the tracking out of the DcpConnMap as it no longer fits as
the owner. The new owner is KVBucket

2) Renames BackfillTrackingIface as KVStoreScanTracker as fundamentally
we care about how many ScanContexts have been created by KVStore::scan
be it RangeScan or DCPBackfill.

3) Modernise the tracking code to remove the std::mutex and instead use
a folly::Synchronised

Future change will extend the KVStoreScanTracker so it counts
RangeScans and prevents the RangeScan create path from exhausting file
resources.

Change-Id: I9e506fecd7ce576aa38e2c59de3c44a08e5e67ad