MB-51979: Add isMetered to Collections::VB::Manifest
This change follows the approach of the scope data-limit in that it
is new metadata that is not replicated (or stored in system-events).
The metered flag is only maintained from the main Manifest.

* A collection created on an active will get the value direct from
the Manifest.

* A collection created via DCP (a replica VB receiving the
create-collection event) will set the value to the default of true
and then detect if that is incorrect if the replica is ever changed
to be active.

Change-Id: I1355eb339a5b6c707821726c51c8814b1707bdfe