MB-51979: Support unmetered collections
Utilise Cookie::currentCollectionInfo to retrieve from the engine
a collection operation's metering state and use that flag in the
code which updates the usage.

Update MeteringTest so that it creates an unmetered collection
and runs a new variant against that collection.

Currently GetRandomKey and RangeScanContinue will fail as they don't
set the metered state on Cookie::currentCollectionInfo

Change-Id: Iebb0432de6a0b8f549e54ade5d3d4549d9533fff