MB-53877: Implement custom move for StoredValueProxy
If a StoredValueProxy is created and moved from, both the objects
will attempt to call "epilogue" from ~StoredValueProxy. We need
a custom move to ensure the moved from object clears the value
member.

This also protects against possible calls into the StoredValue whilst
the StoredValueProxy no longer owns the ht-lock. For this second issue
it was only noticeable with extra debug code such as:

   Expects(lock.owns_lock());

This would fail just because the lock has been moved, not that the lock
itself had been completely released.

Change-Id: I5d388cb2b10b518d4197986a1a34eacd6c137d7b