MB-53877: Release StoredValue when dropping HashTable locks
This commit is the low-hanging fruit to fix as part of the issue
identified in MB-53877 where we have code which can still read
HashTable values after the HashBucketLock has been unlocked.

Rather than just unlocking the HashBucketLock, a release function
now brings the work which would of occurred in the destructor path
into the release path, ensuring all Statistics updates are now safe.

Change-Id: I85a3606b8d5c6e00a65efe4308961a134589c899