MB-54729: Refactor ScanContext to use unsigned type for seqnos
KV does not use negative sequence numbers on in the sequence index, so
a scan should use unsigned types when storing sequence numbers.

RocksDB has 1 special case where a special negative seqno is used in
some code - hence it retains the signed getSeqnoSlice - but that
special case is not related to scans.

Change-Id: Ia05144b2b45926b15aafc6f0e82ab4c94fecde91