MB-: Track the default collection's high legacy DCP seqno
Every collection tracks a high-seqno, and this value includes all
'events' DCP can send. E.g prepare, abort, mutation, system event.

Any 7.0 (collection aware) DCP client and see this high-seqno and
SeqnoAdvance allows DCP to skip any 'events' they don't opt-in to.

Any legacy DCP client (no collections) cannot receive SeqnoAdvance
so KV maintains some extra sequence numbers which can be exposed
to such a client.

1) max-visible-seqno

Added in 7.1, this tracks the highest mutation a vbucket stores (and
which DCP would send). This does not increment for a prepare/abort
or system event.

2) max legacy DCP seqno

New in this commit. A seqno which tracks the highest 'event' a legacy
DCP stream could send (when the right options are enabled). This will
increment for mutation, prepare and abort.

This commit only tracks in-memory and a warmup will lose the value.

Change-Id: Id876e47099380dd90059d98534bc9e0b3b0c9b89