DOUBTS: MB-55451: Default collection MVS incorrect for single prepare
SEE COMMENTS IN vbucket_manifest.cc

The range checking logic added in MB-50413 was incorrect and didn't
account for the case where the default collection consists of only
prepares (mvs should be 0).

The fix is to pass the high-prepared-seqno into the
setDefaultCollectionMaxVisibleSeqnoFromWarmup function and use that
to make an additional decision about the state of the default
collection.

Warmup scans for prepares in the window of:

    high-completed-seqno (hcs) to high-prepare-seqno (hps)

If the default collection's high-seqno is greater then the hps, then
it must be committed and should become the max visible.

If in the window no default collection commit is found (seqno is 0)
and the default collection's high-seqno is less then hps, then the
high-seqno is visible (completed/committed).

ARGH hmm an abort??

Else the given seqno is the max-visible seqno.

warning "still broken??"

writing test for
p@1
a@2 <- hcs, query is can the hcs be above the prepare?
p@3
p@4 <- hps

seqno = 0
highSeqno = 3

Change-Id: Id5e2e00bee9763eb3f4510b92cdb0a2bf2555acc