MB-55672: Bundle-up RangeScan create/continue parameters
The aim of this commit is to reduce the churn of changes required
if a new parameter is required for create/continue. The parameters
are now bundled into structs removing the need to update lots and
lots of class interfaces for future changes.

For MB-55672 a new continue parameter will be added.

Change-Id: I36e83e29679f72cd97d158e60ffc5669b65c9ad8