MB-9418: Tolerate rollback during warm-up
MB-48373 was raised because the scan issued during warm-up returned
a failure, and this was due to a vbucket rolling back to 0.

This only occurred because there was a race with warm-up itself
signalling "complete" which triggered ns_server to create the
consumer. The race was that some backfill tasks were still
scheduled and then able to run.

With secondary warm-up there is now the need to handle the rollback
rather than avoid it, KV cannot defer consumer creation until all
Warmup steps are complete.

This commit adds a unit test which reproduced the original error and
now adds checks to tolerate the failure when the vbucket appears
rolled back.

Change-Id: I9303159131638f1c234f38fe882a4b3d0cf0d70f