MB-41351: VB::Manifest generates a flush event when flush_uid changes
Inject a system-event into a vbucket when the flush_uid of a collection
changes. VB::Manifest is also updated so that it can receive a
flush system-event (ready for DCP integration).

This patch does not yet support a flush or true DCP transmit of the
event, but is tested using existing code from vbucket_manifest_test.cc
which simulates a transfer of an event from active to replica and then
compare of active vs replica.

Change-Id: Idc878cc529a45523e4ed5cb53cb797724e517ed0