Flush Collection compaction stuff
How was this going to be tackled?

Current commit trys to process this by updating the dropCollection state
in flush.cc which triggers some issues elsewhere.

Maybe we just need a flushed flush bool to trigger the compaction

Next the ScanContext thing must always load OpenCollections so it knows
the extent of the purge range.

Change-Id: I9d8353b166c8f58bbfda86134b58226936b0a399

---

Current impl tracks flushes in flush.cc and then adds an entry into the
dropCollections document. This kind of works except for magma as that isn't
doing a vb.itemCOunt update

Change-Id: Ic99ad7557fcb29630e6332f962e87cd5909073a8