MB-63977: Use DCP control to enable purge-seqno tx
Change from using DCP open flags to using DCP control to
enable purge-seqno transmission.

The reason for this is that to more easily permit replication
to utilise purge-seqno transmission to fix MB-63977 we need
to control an existing producer (KV does not control how the
producer is created, ns_server does that). Use of DCP control
in this way is an established pattern for other features
replication wishes to use.

Finally for the orginal MB that added purge-seqno support into
DCP, MB-27800. DCP StreamRequest is changed to send a purge-seqno
to the server using the JSON object and not to use a new framing
of the command.

Thus.

1) To enable a producer to send the purge_seqno, they must request
v2.2 of the DCP marker.

dcp_control("marker_v2_2", "true")

2) Then to avoid rollbacks they can inform the server of their
purge_seqno

dcp_stream_request(..., {"purge_seqno":"x"})

Change-Id: I994ffc6744cbfd0cf05cc276810758a33d481717