MB-63977: Maintain the purge-seqno from flushing
When a purge-seqno is provided (by replication) the vbucket snapshot
must reflect that purge-seqno.

The fix piggy backs upon code managing other "seqnos" which get
replicated and flushed, e.g. the HCS. Checkpoints are given the
replicated purge-seqno which is then picked up by the flusher.

There are few details in the implementation. The replicated purge-seqno
can be in the future of the data arriving, e.g. we may be flushing
items that are below the purge-seqno and in those cases the persisted
purge-seqno never exceeds the end of the flushed range.

Change-Id: I417bf3a3ec9e7cd7c6e1fa43f4d01707823f8456