MB-64123: Permit release_snapshot to use a vbucket as input
To provide a way for the cluster manager to release a snapshot. In
this case the uuid wouldn't be known, or important, the use-case is
to abort a rebalance (or move of a specific vbucket).

release_snapshot(vb) is provided so the cluster manager can force
a snapshot out - which will eventually also best effort interrupt
any ongoing download.

Change-Id: I1624d51bcf23a10d3fc24ce81d0778b2cd2be687