MB-67520: Define item expiry as 32-bit
Documents are given an absolute expiry time and the persisted metadata
only has a 32-bit field for storing that value. Similarly the
hash-table only reserves 32-bits for storing the same. The same is for
deleted documents, where the "delete-time" value is stored in the same
4-bytes.

This commit removes the use of time_t when we are dealing with document
expiry and uses the explicitly sized u32int_t type so it's clear that
32-bits are permitted for expiry.

This then leads to the need for narrowing conversions when we are
dealing with newly generated time stamps, for example the use of
ep_real_time which returns the approximate system time now needs a
narrowing conversion, and in this case to avoid silent dataloss and
expiry bugs we use gsl::narrow to fail if the time ever exceeds u32
storage (which will happen in 2106, but system clocks should not be
trusted certainly if we ever persist data).

Change-Id: I8b7e612dbd890fc621de0df6ef222ab25bfc5677