MB-67520: Change mc_time expiry limit method to use u32
The expiry limiting code exists to take an already calculated item
expiry and check it doesn't exceed some defined limit.

The types are confused in that we are using time_t extensively for
general absolute time (which can and should be the maximum unsigned
int) and document absolute expiry time, which is fundmentally limited
to u32 by the current persisted metadata and less strictly by the
HashTable metadata (currently using u32 for space effciency).

This commit tidies this path up so that a u32int is the input and
output of the lower level server API, a stepping stone to cleaning
up our u32/u64 expiry-time/system-time mismatch and silent narrowing.

Change-Id: I44ec1575bae8b48a1036f27ab806ea3280f32ff5