MB-67520: Add a function for generating document expiry time
Remove duplicated code that converts the protocol provided expiry
into a document expiry time and ensure u32int type is used on the
on the result.

Later this function can be adapted to provide overflow detection in
the case when MCBP relative expiry results in a value that exceeds
u32 storage.

Change-Id: I5bbc539500e96c19503cff6e4347ac0be74939e1