MB-64001: Cancel CacheTransfer if consumer detects memory pressure
When the DcpConsumer steps forwards to process a message it uses
the replication throttle to determine if it should proceed. Here
memory pressure would request the consumer to pause.

If the consumer is paused and the current event is a CacheValue
and this is a full eviction bucket we can stop the cache transfer.
There is no use in pushing more cache items into the consumer when
memory is constrained.

If the bucket is value eviction, that is different because we need
all of the keys, in that case we pause/resume replication just
like regular DCP would.

Change-Id: I0e549faf4a8ab8664dd8ac9e2269d4793036a655