MB-64001: Use a Key/Meta cache-transfer for value eviction
Use a cache-transfer to reach the minimal requirements for a value
eviction vbucket - that is all keys/meta are resisdent.

The value eviction cache-tranfer will attempt to move 100% of the
key/meta from the active vbucket. This leaves us with 0% residency
which can be solved later by a background loading of values.

Change-Id: Iceb7fca83dfd094d1a4e99123c0d714b922228ec