MB-64001: Send and handle DcpCacheTransferEnd
In the producer when we process the switch-to-active message
that is the signal that the cache transfer has ended and we
can send the message.

In the consumer the message is now received and handled but
only logged. Further changes will update some vbucket variable
and expose via a stat group (TBD).

Change-Id: Ie108c2e1ae1677887b1ef952439152a50db8032b