MB-64001: Trigger CacheTransfer once for snapshot created vbuckets
Remove checks of various stats and use

1. A (new) dedicated state variable for determining how a vbucket was
   created.

Warmup, SetVbucket (non snapshot), FBR or Fusion.

2. A one-shot flag to ensure CTS happens no more than once.

CTS is disabled if a vbucket changes state, e.g. we have only considerd
CTS for the initial replica following a FBR/Fusion vbucket move, if the
vbucket flipped to active for a bit and back to replica, we don't want
the CTS to trigger in this case.

Change-Id: I263dd8a1c985beaf3f155cfefec0d0a7d1d75b52